Remodel variable/field declaration AST to one-to-one match the source language (#9)
This patch remodel the AST to exactly match the source program. This helps us get rid of the DeclView workaround and let resolved declaration in AST to directly point to the variable/field declaration. Having an exact match to the source program also helps the language server to preserve more info of the original structure of the program.